Added Music class and fixed a big oopsie in oggplayer
Added class for music playback.
Fixed missing line in oggplayer.c
Noticed that the oggplayer is modified from the one provided with DevkitPro.
Found out that something in menu code or fonts and input code breaks Dolphin in such a way that starting to play music after first frame causes glitches or crash.